+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ### Changes
 
+- ＋ Adds permissions `saito.core.posting.solves.set` for marking a posting as solution/helpful.
 - ＋ Improves compatibility with PHP 7.3
 - ＋ Improves browser detection for changes in the Bota theme CSS
 - ✓ Deleting a bookmark creates an empty area above the bookmarks